完成异常处理

+++ src/main/java/com/cholas/community/controller/QuestionController.java
@@ -18,7 +18,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id, Model model) {
+    public String question(@PathVariable(name = "id") Long id, Model model) {
         QuestionDTO questionDTO = questionService.getById(id);
         //累计阅读数
         questionService.incView(id);
